+++ index.js
@@ -4,8 +4,9 @@ const Log = require('mr-doc-utils').Log;
 const Option = require('mr-doc-utils').Option;
 const Output = require('./src/utils/output');
 const Promise = require('bluebird');
-const Through = require('through2');
 const VinylFS = require('vinyl-fs');
+const Parser = require('./src/parser/');
+const Compiler = require('./src/compiler/');
 const log = new Log();
 
 class MrDoc {
@@ -13,7 +14,7 @@ class MrDoc {
     return new Promise(resolve => {
       const output = options.output || options.o || Option.defaults.mrdoc.output;
       stream
-        .pipe(MrDoc.gulp(options))
+        .pipe(MrDoc.gulp(Option.merge(options, true)))
         .pipe(VinylFS.dest(output))
         .on('end', () => {
           log.debug(Log.color.blue('Mr. Doc compiled successfully'));
@@ -22,13 +23,18 @@ class MrDoc {
     });
   }
   static gulp(options) {
-    const buffer = [];
-    return Through.obj(Output.toBuffer(buffer),
-    Output.handler(buffer, Option.merge(options, true)));
+    return Output.toStream(MrDoc.parser(options), MrDoc.compiler(options), options);
   }
   // static grunt() {
   //   //
   // }
+  static parser(options) {
+    // Initalize Parser and Compiler
+    return (new Parser(options)).factory();
+  }
+  static compiler(options) {
+    return (new Compiler(options)).factory();
+  }
 }
 
 module.exports = MrDoc;
